segment out Sign Up modal
